Refactor the PeepholeOptimizer into arbitrary passes
